EXC_BAD_ACCESS, when adding block to resolved promise
I received an EXC_BAD_ACCESS signal in my iOS7 application, when I tried to send something to the value property. The reason was (at least as far as I understand it) that the value property was released by garbage collection. If you want to reproduce it, try to use the value in a addDone block, which was added after the promise was already resolved. This simple line here should fix it.

P.S.: Thanks for your great library, it made my life so much easier!